Move builder into its own module and break out ServiceBuilderMaker
