+++ tower/src/builder/mod.rs
@@ -1,13 +1,16 @@
 //! Builder types to compose layers and services
 
-use futures::{Async, Future, Poll};
+mod service;
+
+pub use self::service::{MakerFuture, ServiceBuilderMaker};
+
+use never::Never;
 use std::marker::PhantomData;
-use std::sync::Arc;
 use tower_layer::{util::Chain, Layer, LayerExt};
 use tower_service::Service;
 use tower_service_util::MakeService;
 
-use never::Never;
+pub(super) type Error = Box<::std::error::Error + Send + Sync>;
 
 /// Configure and build a `MakeService`
 ///
@@ -133,26 +136,6 @@ pub struct ServiceBuilder<L, S, Request> {
     _pd: PhantomData<(S, Request)>,
 }
 
-/// Composed `MakeService` produced from `ServiceBuilder`
-#[derive(Debug)]
-pub struct ServiceBuilderMaker<S, L, Request> {
-    maker: S,
-    layer: Arc<L>,
-    _pd: PhantomData<Request>,
-}
-
-/// Async resolve the MakeService and wrap it with the layers
-#[derive(Debug)]
-pub struct MakerFuture<S, L, Target, Request>
-where
-    S: MakeService<Target, Request>,
-{
-    inner: S::Future,
-    layer: Arc<L>,
-}
-
-type Error = Box<::std::error::Error + Send + Sync>;
-
 impl<S, Request> ServiceBuilder<Identity, S, Request> {
     /// Create a new `ServiceBuilder` from a `MakeService`.
     pub fn new() -> Self {
@@ -182,11 +165,7 @@ impl<L, S, Request> ServiceBuilder<L, S, Request> {
         M: MakeService<Target, Request, Service = S, Response = S::Response, Error = S::Error>,
         S: Service<Request>,
     {
-        ServiceBuilderMaker {
-            maker,
-            layer: Arc::new(self.layer),
-            _pd: PhantomData,
-        }
+        ServiceBuilderMaker::new(maker, self.layer)
     }
 
     /// Wrap the service `S` with the layers.
@@ -196,50 +175,6 @@ impl<L, S, Request> ServiceBuilder<L, S, Request> {
         S: Service<Request>,
     {
         self.layer.layer(service)
-    }
-}
-
-impl<S, L, Target, Request> Service<Target> for ServiceBuilderMaker<S, L, Request>
-where
-    S: MakeService<Target, Request>,
-    S::MakeError: Into<Error>,
-    L: Layer<S::Service, Request> + Sync + Send + 'static,
-    L::LayerError: Into<Error>,
-    Target: Clone,
-{
-    type Response = L::Service;
-    type Error = Error;
-    type Future = MakerFuture<S, L, Target, Request>;
-
-    fn poll_ready(&mut self) -> Poll<(), Self::Error> {
-        self.maker.poll_ready().map_err(Into::into)
-    }
-
-    fn call(&mut self, target: Target) -> Self::Future {
-        let inner = self.maker.make_service(target);
-        let layer = Arc::clone(&self.layer);
-
-        MakerFuture { inner, layer }
-    }
-}
-
-impl<S, L, Target, Request> Future for MakerFuture<S, L, Target, Request>
-where
-    S: MakeService<Target, Request>,
-    S::MakeError: Into<Error>,
-    L: Layer<S::Service, Request>,
-    L::LayerError: Into<Error>,
-{
-    type Item = L::Service;
-    type Error = Error;
-
-    fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
-        let service = try_ready!(self.inner.poll().map_err(Into::into));
-
-        match self.layer.layer(service) {
-            Ok(service) => Ok(Async::Ready(service)),
-            Err(e) => Err(e.into()),
-        }
     }
 }
